\name{IGRF13-package}
\alias{IGRF13-package}
\alias{IGRF13}
\docType{package}
\title{
International Geomagnetic Reference Field
}
\description{
R-interface to International Geomagnetic Reference Field version 13  FORTRAN routines
}
\details{
\tabular{ll}{
Package: \tab IGRF13\cr
Version: \tab 0.1-0\cr
Date: \tab 2020-05-11\cr
License: \tab \cr
Depends: \tab \cr
}

Index:

\code{\link{igrf}} The model function in R


}
\author{
  Ilkka Virtanen (University of Oulu, Finland) \cr
  \email{ilkka.i.virtanen@oulu.fi}
}
\keyword{ package }
\references{The package uses the FORTRAN routine igrf13syn from the
  package by Susan Macmillan and William Brown http://www.ngdc.noaa.gov/IAGA/vmod/igrf.html}